pyproject.toml
[project]
name = "engineai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "channels>=4.3.2",
    "django-ai-assistant>=0.1.2",
    "litellm>=1.80.0",
    "openai>=1.109.1",
]
